tests/funding/requirements.txt
azure-identity==1.20.0
azure-keyvault-secrets==4.9.0
openai==1.55.0
requests==2.32.3
tqdm==4.67.1
